+++ offlinemsmtp/daemon.py
@@ -1,5 +1,8 @@
 import os
+import re
+import socket
 import time
+
 from queue import Queue
 from subprocess import PIPE, run
 
@@ -26,8 +29,8 @@ class Daemon(FileSystemEventHandler):
             self.queue.put(os.path.join(args.dir, file))
 
     def send_enabled(self):
-        return (self.send_mail_file is None
-                or os.path.exists(self.send_mail_file))
+        return (
+            self.send_mail_file is None or os.path.exists(self.send_mail_file))
 
     def on_created(self, event):
         """Handle file creation."""
@@ -54,13 +57,7 @@ class Daemon(FileSystemEventHandler):
                 msmtp_args = message_content.readline().decode()
                 message_content = message_content.read()
 
-            account = 'ohea' # TODO
-
-            if not util.test_internet(self.config_file, account):
-                util.notify(
-                    f'Cannot connect to SMTP server to send: {message}',
-                    timeout=5000,
-                )
+            if not self.can_send_message(msmtp_args, message_content):
                 failed.append(message)
                 continue
 
@@ -72,11 +69,8 @@ class Daemon(FileSystemEventHandler):
             )
 
             # Send the message.
-            command = [
-                '/usr/bin/msmtp', '-C', self.config_file, *msmtp_args.split()
-            ]
             sender = run(
-                command,
+                self.get_msmtp_command(msmtp_args),
                 input=message_content,
                 stdout=PIPE,
                 stderr=PIPE,
@@ -100,6 +94,63 @@ class Daemon(FileSystemEventHandler):
         # Re-enqueue the failed messages.
         for f in failed:
             self.queue.put(f)
+
+    host_re = re.compile('host = (.*)')
+    port_re = re.compile('port = (.*)')
+    subject_re = re.compile('Subject: (.*)')
+
+    def get_msmtp_command(self, msmtp_args, pretend=False):
+        args = ['/usr/bin/msmtp']
+        if pretend:
+            args.append('-P')
+        args += ['-C', self.config_file, *msmtp_args.split()]
+        return args
+
+    def can_send_message(self, msmtp_args, message_content):
+        """
+        Tests whether or not the computer can connect to the necessary server
+        to send the given message.
+        """
+        test_run = run(
+            self.get_msmtp_command(msmtp_args, pretend=True),
+            input=message_content,
+            stdout=PIPE,
+            stderr=PIPE,
+        )
+
+        for line in test_run.stdout.decode('utf-8').split('\n'):
+            # TODO conver this stuff to Walrus operators once that's more
+            # supported by tools.
+            host_match = self.host_re.match(line)
+            if host_match:
+                host = host_match.group(1)
+                continue
+
+            port_match = self.port_re.match(line)
+            if port_match:
+                port = int(port_match.group(1))
+
+        # Try to connect to the socket.
+        sock = socket.socket()
+        sock.settimeout(2)  # 2 second timeout
+        socket_open = sock.connect_ex((host, port))
+        sock.close()
+
+        # Notify if it's not available.
+        if socket_open != 0:
+            # Search for the subject in the message_content
+            subject = '<no subject>'
+            for line in message_content.decode('utf-8').split('\n'):
+                subject_match = self.subject_re.match(line)
+                if subject_match:
+                    subject = subject_match.group(1)
+
+            util.notify(
+                f'Cannot connect to {host}:{port} to send message with '
+                f'subject: "{subject}".',
+                timeout=5000,
+            )
+        return socket_open == 0
 
     @staticmethod
     def run(args):