+++ CHANGELOG.rst
@@ -1,7 +1,9 @@
 v0.3.5
 ======
 
-* 
+* Use a real socket to try and connect to the SMTP server instead of ping.
+* ``offlinemsmtp`` only tries to connect to the server that it is sending mail
+  to for determining if it should attempt to send that element of the queue.
 
 v0.3.4
 ======